touched up some more
